ENH: basic @algebraic macro
In order to make symbolic ODEs, we first have to have symbolic functions.

This commit is the first step towards that, introducing the `@algebraic` macro
which makes a DSL for introducing symbolic functions.

These hook into the Julia scope system, so they can conveniently call
other symbolic functions which happen to be in scope.