more principled approach to empty tries
We want a trie to be uniquely determined by the set of valid TrieVars
that it supports. If we allow there to be empty branches, i.e. corresponding
to the named tuple `(a::(b::()))`, then this has no valid paths in it,
but is different from `()`. To solve this, we make two types: `Trie`
and `NonEmptyTrie`. The recursive case is `NonEmptyTrie`: a `Trie`
is just a wrapper around `Union{Nothing, NonEmptyTrie{A}}`.

This was inspired by the way yuujinchou works.

Also new in this commit is the use of MLStyle active patterns to
make pattern matching on tries cleaner. This is a huge upgrade to the
previous way we were doing things, and I think that we should use more
active patterns in the future throughout GATlab, for pattern matching
on unityped data. See the active patterns Leaf, Node, Empty, Root, Nested
in src/syntax/Tries.jl for more information.